Rewrite of the whole program (one feature still missing!)
Many functions were removed by inlining them into their calling
function. Unnecessary function splitting has lead to many small
functions that don't need to be seperated. As a result the code
should be shorter and less spammed with function declarations.

A few different things were done, too:
  - the regexes for ignored filenames are not compiled every
    time a file changes anymore
  - some expressions were simplified to shrink the code
  - some things were renamed to match a more idiomatic
    naming scheme
  - comments were added

Important: Currently the feature of waiting 2 seconds before
starting a new compile is removed! It is planned to be replaced
by a better way of scheduling. However, this was not yet
implemented. This can lead to excessive process spawning when
many files change in a short time. Do not use in production!