When starting, do a command run before any file changes
I've often started cargo watch when started working on a project, and
it clearly doesn't do anything until a file has changed. Sometimes I
can't remember if all the tests passed (or if the code compiles), and
I'd like to see the output of cargo test. So I need to make a simple
change to something so that it'll run and I can see. But if cargo watch
ran the command as soon as it was started, this would be easier.

fixes #37