[watch] Implement own handler wrapping watchexec’s
With our own handler, we get more control over the starting of commands
(alas, watchexec doesn't keep track of when a command finishes, so we
are still bound to alter the command for the finish announce).

That lets us eliminate the "Running..." announce entirely from the
command, and print it out from Rust, which means no more quoting issues
(in this regard)! The finish announce still needs to be in the command,
see above, but given it contains no user input at all, the quoting can
also be eliminated.

Additionally, this commit changes the joining of that announce to the
rest, using `;` rather than `&&`, which means the announce will be
printed regardless of whether the command is successful.

And right off that improvement, on Unix we use a slightly different
announce that includes `$?`... showing the exit status.