Use caretaker maintainership
I've thought about this for a moment. I have been hesitant to apply
caretaker maintainership to cargo-watch even though the impetus for
creating this caretaker concept was the whole situation with cargo-watch
and notify and co. Small projects are a different beast to something
that is used throughout the community and has 1.5k stars on github, as
much or as little you can attach any value to such measurement.

Ultimately, though, this is the same kind of thinking that got me to so
spectacularly and decisively burn out of notify. The truth of the size
of the project is that it's mostly just me. Just me writing most of the
code, just me doing the releases, just me managing issues and requests
that have been piling on without resolution for the best part of three
years.

Caretaker, at its heart, is merely a formalisation and acceleration of a
process that already happens in open source. It does that to try to make
it even easier and more common, yes, but it's not fundamentally creating
anything new.

IF, AS A RESULT OF THIS CHANGE, YOU HAVE CONCERNS ABOUT SUPPLY CHAIN
ATTACKS, YOUR ONLY WAY TO SOLVE IT IS TO ***ACTIVELY*** TAKE OVER THE
PROJECT AS A COMMITTED MAINTAINER. If you are truly committed, I will
even facilitate this as best I can.

Kia ora