Run the tests (in CI) only on Linux
That’s historical, see for eg the previous Travis config:
https://github.com/passcod/cargo-watch/blob/v7.3.0/.travis.yml

From what I remember the idea is the tests we have are mostly cosmetic
and not functional, and it all varies too much on different platforms to
matter much, so they’re there as a “make sure you haven’t fucked up”
last chance catch. The actual tests are in watchexec.