Support --stdin-quit functionality
The core behaviour has been added to the watchexec lib so we update to
the latest watchexec and add the flag & hook into the runtime to get
the desired behaviour.

The switch to a git sha rev for watchexec should be temporary.

The other package updates seemed to be required and come from deleting
the Cargo.lock entries for once_cell and thiserror to allow them to
update and then updating the version of miette to match the one used
by watchexec.